feat: remove 'employee' from TimePeriod to avoid redundant information
The endpoint to return the time-periods is always employee-specific so
we don't need this redundant information there.

Conceptually it would make sense to make timePeriods a fields of the
Employee class but since we are using Employee also for the payroll part
of the application we want to separate those concerns. Tricky question.