tests: add test to validate uknown employee returns no work periods
This is a behavior that I probably want to change later, instead
returning an error, but for know I add the test to document the
expectation. Returning an empty list has the advantage of avoiding
leaking information about what employee IDs are valid and which are not,
but since in real life we would want to add authorization to access only
certain employees (in particular your own data) this would no longer
make sense.