fix: H2 database problem with reserved keywords
This was leading to some very cryptic error when creating a new Entity
class in a followup commit.

Basically columns (and tables) can't be named using the reserved
keywords, which are many! For instance "month" is one of them. A way
around this if these words are quoted, which is what the new option
ensures.

Reserved keywords: http://www.h2database.com/html/advanced.html#keywords

Similar problems:

* https://github.com/spring-projects/spring-data-r2dbc/issues/800
* https://www.baeldung.com/spring-boot-h2-jdbcsqlsyntaxerrorexception-expected-identifier